+++ src/neo-vm/Types/ByteString.cs
@@ -1,3 +1,4 @@
+using Neo.VM.Cryptography;
 using System;
 using System.Buffers.Binary;
 using System.Diagnostics;
